partial cached subtree detection
cached subtree consists in detecting that a subtree has already
been appeared in the grammar.  We implement them to appear to
detect repeated substructures like same alternatives rules or
alternative lists, repeated last element of a repeating scheme
in front of it, to allow for not nullable lists of elements.

We plan to identify sequences of the type:

A B { C D }

identify the sequence C D on it, and substitute it with

list_C_D: list_C_D C D
        | A B ;

Signed-off-by: Luis Colorado <luiscoloradourcola@gmail.com>