Restored old syntax to match Wirth's 1980 spec.
Finally I refuse to extend the original syntax from N. Wirth, and
just modify the DefinitionModule syntax not to include the final period
and put it in the CompilationUnit, as with implementation module
and module syntaxes.  This way, if in the future I decide to
extend it, it would be easier.  No difference in the grammar should
be expected from this modification, as DefinitionModule symbol is only
used in CompilationUnit.

Signed-off-by: Luis Colorado <luiscoloradourcola@gmail.com>