define nodes of parse tree.
nodes are defined based on a flexible structure, as opposed
to define a hundred ones with different fields, we'll use a
variable size structure with space for an array of descendants
for each node.  Also a pointer to the parent and the node
description is included in the data associated with each tree
node.

Signed-off-by: Luis Colorado <luiscoloradourcola@gmail.com>